reminder-bot/requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-dateutil>=2.8.2
